# backend/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/api/users/me":{"get":{"tags":["user-controller"],"operationId":"getMe","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/bandeiras/{id}/members/role":{"put":{"tags":["bandeira-controller"],"operationId":"updateMemberRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/runs":{"get":{"tags":["run-controller"],"operationId":"getMyRuns","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RunResponse"}}}}}}},"post":{"tags":["run-ingestion-controller"],"operationId":"submitRunWithGpx","parameters":[{"name":"origin","in":"query","required":false,"schema":{"type":"string","default":"IMPORT","enum":["IOS","WEB","IMPORT"]}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}},"application/json":{"schema":{"$ref":"#/components/schemas/SubmitRunRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/RunSubmissionResponse"},{}]}}}}}}},"/api/runs/coordinates":{"post":{"tags":["run-controller"],"operationId":"submitRunWithCoordinates","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitCoordinatesRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RunSubmissionResponse"}}}}}}},"/runs/coordinates":{"post":{"tags":["run-controller"],"operationId":"submitRunWithCoordinates_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitCoordinatesRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RunSubmissionResponse"}}}}}}},"/api/runs":{"get":{"tags":["run-controller"],"operationId":"getMyRuns_1","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RunResponse"}}}}}}},"post":{"tags":["run-controller"],"operationId":"submitRunWithGpx_1","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RunSubmissionResponse"}}}}}}},"/api/bandeiras":{"get":{"tags":["bandeira-controller"],"operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BandeiraDto"}}}}}}},"post":{"tags":["bandeira-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBandeiraRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BandeiraDto"}}}}}}},"/api/bandeiras/{id}/join":{"post":{"tags":["bandeira-controller"],"operationId":"join","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BandeiraDto"}}}}}}},"/api/bandeiras/leave":{"post":{"tags":["bandeira-controller"],"operationId":"leave","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/api/auth/register":{"post":{"tags":["user-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/signup":{"post":{"tags":["user-controller"],"operationId":"signup_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/refresh":{"post":{"tags":["user-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["user-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/me":{"get":{"tags":["user-controller"],"operationId":"getMe_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/runs/{id}":{"get":{"tags":["run-controller"],"operationId":"getRunById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RunResponse"}}}}}}},"/runs/{id}":{"get":{"tags":["run-controller"],"operationId":"getRunById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RunResponse"}}}}}}},"/runs/daily-status":{"get":{"tags":["run-controller"],"operationId":"getDailyStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DailyStatusResponse"}}}}}}},"/api/runs/daily-status":{"get":{"tags":["run-controller"],"operationId":"getDailyStatus_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DailyStatusResponse"}}}}}}},"/api/quadras":{"get":{"tags":["quadra-controller"],"operationId":"getQuadras","parameters":[{"name":"minLat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"minLng","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"maxLat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"maxLng","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"centerLat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"centerLng","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"radiusMeters","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"bbox","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"oneOf":[{"type":"array","items":{"$ref":"#/components/schemas/ViewportQuadraDto"}},{"type":"array","items":{"$ref":"#/components/schemas/QuadraDto"}}]}}}}}}},"/api/quadras/{id}":{"get":{"tags":["quadra-controller"],"operationId":"getQuadraById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuadraDto"}}}}}}},"/api/quadras/user/{userId}":{"get":{"tags":["quadra-controller"],"operationId":"getQuadrasByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuadraDto"}}}}}}}},"/api/quadras/stats":{"get":{"tags":["quadra-controller"],"operationId":"getStats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuadraStats"}}}}}}},"/api/quadras/disputed":{"get":{"tags":["quadra-controller"],"operationId":"getDisputedQuadras","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuadraDto"}}}}}}}},"/api/quadras/bandeira/{bandeiraId}":{"get":{"tags":["quadra-controller"],"operationId":"getQuadrasByBandeira","parameters":[{"name":"bandeiraId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuadraDto"}}}}}}}},"/api/quadras/at":{"get":{"tags":["quadra-controller"],"operationId":"getQuadraAtCoordinate","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"lng","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuadraDto"}}}}}}},"/api/bandeiras/{id}":{"get":{"tags":["bandeira-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BandeiraDto"}}}}}}},"/api/bandeiras/{id}/members":{"get":{"tags":["bandeira-controller"],"operationId":"getMembers","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BandeiraMemberDto"}}}}}}}},"/api/bandeiras/search":{"get":{"tags":["bandeira-controller"],"operationId":"search","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BandeiraDto"}}}}}}}},"/api/bandeiras/rankings":{"get":{"tags":["bandeira-controller"],"operationId":"getRankings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BandeiraDto"}}}}}}}},"/api/admin/telemetry/runs":{"get":{"tags":["run-telemetry-controller"],"operationId":"exportRuns","parameters":[{"name":"format","in":"query","required":false,"schema":{"type":"string","default":"json"}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UpdateProfileRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":30,"minLength":3},"avatarUrl":{"type":"string"},"isPublic":{"type":"boolean"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"username":{"type":"string"},"avatarUrl":{"type":"string"},"isPublic":{"type":"boolean"},"bandeiraId":{"type":"string","format":"uuid"},"bandeiraName":{"type":"string"},"role":{"type":"string"},"totalRuns":{"type":"integer","format":"int32"},"totalDistance":{"type":"number","format":"double"},"totalDistanceMeters":{"type":"number","format":"double"},"totalTilesConquered":{"type":"integer","format":"int32"}},"required":["email","id","isPublic","role","totalDistance","totalDistanceMeters","totalRuns","totalTilesConquered","username"]},"UpdateRoleRequest":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"role":{"type":"string"}},"required":["role","userId"]},"CoordinatePoint":{"type":"object","properties":{"lat":{"type":"number","format":"double"},"lng":{"type":"number","format":"double"}},"required":["lat","lng"]},"SubmitRunRequest":{"type":"object","properties":{"coordinates":{"type":"array","items":{"$ref":"#/components/schemas/CoordinatePoint"}},"timestamps":{"type":"array","items":{"type":"integer","format":"int64"}},"origin":{"type":"string","enum":["IOS","WEB","IMPORT"]}},"required":["coordinates","timestamps"]},"SubmitCoordinatesRequest":{"type":"object","properties":{"coordinates":{"type":"array","items":{"$ref":"#/components/schemas/CoordinatePoint"}},"timestamps":{"type":"array","items":{"type":"integer","format":"int64"}},"mode":{"type":"string","enum":["COMPETITIVE","TRAINING"]},"targetQuadraId":{"type":"string"}},"required":["coordinates","mode","timestamps"]},"CapsRemaining":{"type":"object","properties":{"userActionsRemaining":{"type":"integer","format":"int32"},"bandeiraActionsRemaining":{"type":"integer","format":"int32"}},"required":["userActionsRemaining"]},"LoopValidationResponse":{"type":"object","properties":{"isValid":{"type":"boolean"},"distance":{"type":"number","format":"double"},"duration":{"type":"integer","format":"int32"},"closingDistance":{"type":"number","format":"double"},"quadrasCovered":{"type":"array","items":{"type":"string"}},"primaryQuadra":{"type":"string"},"primaryQuadraCoverage":{"type":"number","format":"double"},"fraudFlags":{"type":"array","items":{"type":"string"}},"failureReasons":{"type":"array","items":{"type":"string"}}},"required":["closingDistance","distance","duration","failureReasons","fraudFlags","isValid","primaryQuadraCoverage","quadrasCovered"]},"OwnerSnapshot":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["SOLO","BANDEIRA"]}}},"RunResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"origin":{"type":"string","enum":["IOS","WEB","IMPORT"]},"status":{"type":"string","enum":["RECEIVED","VALIDATED","REJECTED"]},"distance":{"type":"number","format":"double"},"distanceMeters":{"type":"number","format":"double"},"duration":{"type":"integer","format":"int32"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"minLat":{"type":"number","format":"double"},"minLng":{"type":"number","format":"double"},"maxLat":{"type":"number","format":"double"},"maxLng":{"type":"number","format":"double"},"isLoopValid":{"type":"boolean"},"loopDistance":{"type":"number","format":"double"},"loopDistanceMeters":{"type":"number","format":"double"},"territoryAction":{"type":"string"},"targetQuadraId":{"type":"string"},"isValidForTerritory":{"type":"boolean"},"fraudFlags":{"type":"array","items":{"type":"string"}},"createdAt":{"type":"string","format":"date-time"}},"required":["createdAt","distance","distanceMeters","duration","endTime","fraudFlags","id","isLoopValid","isValidForTerritory","origin","startTime","status","userId"]},"RunSubmissionResponse":{"type":"object","properties":{"run":{"$ref":"#/components/schemas/RunResponse"},"loopValidation":{"$ref":"#/components/schemas/LoopValidationResponse"},"territoryResult":{"$ref":"#/components/schemas/TerritoryResultResponse"},"turnResult":{"$ref":"#/components/schemas/TurnResult"},"dailyActionsRemaining":{"type":"integer","format":"int32"}},"required":["dailyActionsRemaining","loopValidation","run","turnResult"]},"TerritoryResultResponse":{"type":"object","properties":{"success":{"type":"boolean"},"actionType":{"type":"string"},"reason":{"type":"string"},"ownerChanged":{"type":"boolean"},"shieldChange":{"type":"integer","format":"int32"},"shieldBefore":{"type":"integer","format":"int32"},"shieldAfter":{"type":"integer","format":"int32"},"inDispute":{"type":"boolean"},"quadraId":{"type":"string"}},"required":["inDispute","ownerChanged","shieldAfter","shieldBefore","shieldChange","success"]},"TurnResult":{"type":"object","properties":{"actionType":{"type":"string","enum":["CONQUEST","ATTACK","DEFENSE"]},"quadraId":{"type":"string"},"h3Index":{"type":"string"},"previousOwner":{"$ref":"#/components/schemas/OwnerSnapshot"},"newOwner":{"$ref":"#/components/schemas/OwnerSnapshot"},"shieldBefore":{"type":"integer","format":"int32"},"shieldAfter":{"type":"integer","format":"int32"},"cooldownUntil":{"type":"string","format":"date-time"},"disputeState":{"type":"string","enum":["NONE","STABLE","DISPUTED"]},"capsRemaining":{"$ref":"#/components/schemas/CapsRemaining"},"reasons":{"type":"array","items":{"type":"string"}}},"required":["capsRemaining","reasons"]},"CreateBandeiraRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":50,"minLength":3},"category":{"type":"string"},"color":{"type":"string","pattern":"^#[0-9A-Fa-f]{6}$"},"description":{"type":"string","maxLength":500,"minLength":0}},"required":["category","color","name"]},"BandeiraDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"slug":{"type":"string"},"category":{"type":"string"},"color":{"type":"string"},"logoUrl":{"type":"string"},"description":{"type":"string"},"memberCount":{"type":"integer","format":"int32"},"totalTiles":{"type":"integer","format":"int32"},"createdById":{"type":"string","format":"uuid"},"createdByUsername":{"type":"string"}},"required":["category","color","createdById","createdByUsername","id","memberCount","name","slug","totalTiles"]},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string","maxLength":30,"minLength":3},"password":{"type":"string","maxLength":2147483647,"minLength":6}},"required":["email","password","username"]},"AuthResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserResponse"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken","user"]},"RefreshRequest":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"LogoutRequest":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"DailyStatusResponse":{"type":"object","properties":{"userActionsUsed":{"type":"integer","format":"int32"},"userActionsRemaining":{"type":"integer","format":"int32"},"bandeiraActionsUsed":{"type":"integer","format":"int32"},"bandeiraActionCap":{"type":"integer","format":"int32"}},"required":["userActionsRemaining","userActionsUsed"]},"QuadraDto":{"type":"object","properties":{"id":{"type":"string"},"lat":{"type":"number","format":"double"},"lng":{"type":"number","format":"double"},"boundary":{"type":"array","items":{"type":"array","items":{"type":"number","format":"double"}}},"ownerType":{"type":"string"},"ownerId":{"type":"string","format":"uuid"},"ownerName":{"type":"string"},"ownerColor":{"type":"string"},"shield":{"type":"integer","format":"int32"},"isInCooldown":{"type":"boolean"},"isInDispute":{"type":"boolean"},"championUserId":{"type":"string","format":"uuid"},"championBandeiraId":{"type":"string","format":"uuid"},"championName":{"type":"string"},"guardianId":{"type":"string","format":"uuid"},"guardianName":{"type":"string"}},"required":["boundary","id","isInCooldown","isInDispute","lat","lng","shield"]},"ViewportQuadraDto":{"type":"object","properties":{"h3Index":{"type":"string"},"ownerType":{"type":"string"},"ownerId":{"type":"string","format":"uuid"},"shield":{"type":"integer","format":"int32"},"dispute":{"type":"boolean"},"cooldownUntil":{"type":"string","format":"date-time"},"colorKey":{"type":"string"}},"required":["dispute","h3Index","shield"]},"QuadraStats":{"type":"object","properties":{"totalTiles":{"type":"integer","format":"int32"},"ownedTiles":{"type":"integer","format":"int32"},"neutralTiles":{"type":"integer","format":"int32"},"tilesInDispute":{"type":"integer","format":"int32"},"disputePercentage":{"type":"integer","format":"int32"}},"required":["disputePercentage","neutralTiles","ownedTiles","tilesInDispute","totalTiles"]},"BandeiraMemberDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"avatarUrl":{"type":"string"},"role":{"type":"string"},"totalTilesConquered":{"type":"integer","format":"int32"}},"required":["id","role","totalTilesConquered","username"]}}}}